Rename utils::rustc module to utils::rustc_version
